fix(agentd): avoid u32 overflow when dropping high capability bounding-set bits
The loop iterates capabilities 0..=63, but 1u32 << 31 already overflows in

debug builds and high bits were never tested correctly. Cast the keep mask to

u64 and shift a u64 so every capability slot is addressed without panic.

Refs: #2101

+++ crates/mvm-agentd/src/guest_mount.rs
@@ -618,7 +618,7 @@ fn raise_ambient_capabilities(capabilities: u32) -> std::io::Result<()> {
 #[cfg(target_os = "linux")]
 fn drop_capability_bounding_set_to(keep: u32) -> std::io::Result<()> {
     for cap in 0..=63u32 {
-        if keep & (1u32 << cap) != 0 {
+        if keep as u64 & (1u64 << cap) != 0 {
             continue;
         }
         let rc = unsafe { libc::prctl(PR_CAPBSET_DROP, cap as libc::c_ulong, 0, 0, 0) };
